Gameplay Settings
Added basic settings for during gameplay. View sensitivity is obvious.
Linearity: Control forces are capped at 1. If linearity is all the way to the right, everytime you press a key, full forces are applied. If not all the way to the right, that amount of force is added to the forces every frame. Ex. a value of 0.5 means it takes two frames to get full forces applied.